+++ src/components/FxBarrel.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import type { FxEmits } from '../lib/emits'
 import { inject, onUnmounted } from 'vue'
-import initGameObject from '../lib/initGameObject'
+import { initFilter } from '../lib/initComponent'
 import commonProps from '../lib/props'
 import { InjectionKeys } from '../lib/provider'
 
@@ -21,7 +21,7 @@ if (!fxController) {
   throw new Error(`filters.${props.external ? 'external' : 'internal'} is not available. Make sure the game object supports filters and WebGL renderer is enabled.`)
 }
 const barrel = fxController.addBarrel(props.amount)
-initGameObject(barrel, props, { isFx: true })
+initFilter(barrel, props)
 onUnmounted(() => {
   if (gameObject.filters)
     fxController.remove(barrel)
