+++ src/components/FxColorMatrix.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import type { FxEmits } from '../lib/emits'
 import { inject, onUnmounted } from 'vue'
-import initGameObject from '../lib/initGameObject'
+import { initFilter } from '../lib/initComponent'
 import commonProps from '../lib/props'
 import { InjectionKeys } from '../lib/provider'
 
@@ -38,7 +38,7 @@ if (!fxController) {
   throw new Error(`filters.${props.external ? 'external' : 'internal'} is not available. Make sure the game object supports filters and WebGL renderer is enabled.`)
 }
 const colorMatrix = fxController.addColorMatrix()
-initGameObject(colorMatrix, props, { isFx: true })
+initFilter(colorMatrix, props)
 onUnmounted(() => {
   if (gameObject.filters)
     fxController.remove(colorMatrix)
