+++ src/components/FxWipe.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import type { FxEmits } from '../lib/emits'
 import { inject, onUnmounted } from 'vue'
-import initGameObject from '../lib/initGameObject'
+import { initFilter } from '../lib/initComponent'
 import commonProps from '../lib/props'
 import { InjectionKeys } from '../lib/provider'
 
@@ -25,7 +25,7 @@ if (!fxController) {
   throw new Error(`filters.${props.external ? 'external' : 'internal'} is not available. Make sure the game object supports filters and WebGL renderer is enabled.`)
 }
 const wipe = fxController.addWipe(props.wipeWidth, props.direction, props.axis, props.reveal)
-initGameObject(wipe, props, { isFx: true })
+initFilter(wipe, props)
 onUnmounted(() => {
   if (gameObject.filters)
     fxController.remove(wipe)
