+++ src/components/TilemapLayer.vue
@@ -3,7 +3,7 @@ import type { PropType } from 'vue'
 import type { CreateOnlyEmits } from '../lib/emits'
 import * as Phaser from 'phaser'
 import { inject } from 'vue'
-import initGameObject from '../lib/initGameObject'
+import { initGameObject } from '../lib/initComponent'
 import commonProps, { gameObjectProps } from '../lib/props'
 import { InjectionKeys } from '../lib/provider'
